Crate hyperlocal_with_windows

hyperlocal-with-windows provides Hyper bindings for Unix domain sockets.

See the UnixClientExt docs for how to configure clients.

See the UnixServerExt docs for how to configure servers.

The UnixConnector can be used in the [hyper::Client] builder interface, if required.

Features

Client- enables the client extension trait and connector. Enabled by default.

Server- enables the server extension trait. Enabled by default.

Structs

SocketIncoming

A stream of connections from binding to a socket.

UnixConnector

the [UnixConnector] can be used to construct a [hyper::Client] which can speak to a unix domain socket.

Uri

A convenience type that can be used to construct Unix Domain Socket URIs

Traits

UnixClientExt

Extention trait for constructing a hyper HTTP client over a Unix domain socket.

UnixServerExt

Extension trait for provisioning a hyper HTTP server over a Unix domain socket.