Function hyperlocal_with_windows::remove_unix_socket_if_present

pub async fn remove_unix_socket_if_present(path: impl AsRef<Path>) -> Result<()>

Helper function to delete a Unix socket if it exists. This function correctly handles Unix sockets on Windows which need to be handled specially.

Errors

Refer to tokio::fs::remove_file.